remove: 스키마, 컴포넌트 정리
